requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.3.0
django-extensions>=3.2.3

# Create requirements.txt for bot framework emulator
botbuilder-core
aiohttp

# Slack Integration
slack-sdk>=3.19.5
slack-bolt>=1.16.1

# Note: asyncio should not be installed from PyPI as it's part of the standard library
# and causes syntax errors with Python 3.11
